Added fish data parsing
Added initial data grouping and normalization for eventual writing to csv file.

The headers will be the first 14 entries of the parsed data. The first column (image) is not yet collected and might be serialized to this field eventually.

The data is then normalized by trimming whitespace, and stripping out any of the last optional column that detail recipes, quests or gifts. This gives us equal chunks of 21 that we can then parse into separate lists.

Once in these lists, the prices need to be grouped in 3 groups of 4, as they represent the different quality tier prices when no perk, the angler perk, or the fisher perk is applied.

The next step would be to write this to collect image data and write to files.