WIP - Testing out better html parsing
BS4 is way more robust than I initially knew, leading to the first arduous way of parsing items from the HTML. This is currently an expirament I'm running to see is BS4 parsing is clearner (which I bet it is). If this is resolved to be better, than I will replace the previous method with this.