WIP - Added HTML Caching and Header and body parsing
Added html retrieval serialization and retrieval to prevent wiki scraping every time I need to test the application. This will need to be cleaned up and broken out into other methods, but the proof of concept seems to be working. Logging will also need to be added in the clean up.

The header parsing was refined and can now be done all through bs4. The string is retrieved, converted to str through an f-string, and then trimmed. The reason f-string is used over .format() is for performance and cleanliness. See: [f-string vs format](https://medium.com/@DahlitzF/pythons-f-strings-vs-str-e22995cefef6)